Move `h.util.cors` => `h.views.api.helpers.cors`
